Folia Support
This should add Folia to the plugin, will need a better VersionUtils check for later as more Folia server versions come out (Added TODO: to each instance that VersionUtils ran so people will know where), currently set to work for 1.20.4 Folia.

As a small overview of what I did:
- Switched all schedules to the UniversalScheduler ones to support Folia (Could implement your own if you want less dependencies).
- Added "folia-supported: true" to the plugin.yml file so it doesn't auto-shutdown.
- Made a quick check if it's folia to use the 1.20.4 version (since it works generally) on all checks for VersionUtils.

NOTE: I haven't fully tested this out, but from the looks of it, there are no errors on boot and the detection of server versions does work, will need some actual testing to see if it works properly.